requirements.txt
Django==2.2.4
django-sortedm2m==3.0.2
